requirements.txt
aiogram>=3.4.1
yt-dlp>=2024.1.1
aiohttp>=3.9.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
